+++ packages/core/src/cli-commands/create-post.js
@@ -61,6 +61,7 @@ export async function createPost(args) {
   const content = `---
 time: ${time}
 slug: ${slug}
+draft: false
 title: ${title}
 tags:
 ${tagsBlock}